requirements.txt
azureml-core==1.58.0
firebase-admin==6.5.0
mlflow==2.17.0
numpy==1.26.4
pandas==2.2.3
plotly==5.24.1
pydeck==0.9.1
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
streamlit==1.39.0
tensorflow==2.17.0
xgboost==2.1.1
uvicorn==0.32.0
fastapi==0.115.4
gunicorn==23.0.0